Make sampler function tunable, but only for substatters.
A tunable substatter allows for pre-sampling of metrics -- where a rate
is simply passed along as the rate value, but no sampling is performed.
see: issue #24

Allowing this only on a substatter (interface) so as to help prevent
accidental misuse by plugins of larger applications -- substatter
changes will not modify statter for all usages.

This will also allow parts of a codebase to create a "pre-sampled"
substatter (with optional stat suffix), for its own internal use.